refactor(WinningLottoNumbers): WinningStatistics 를 반환하도록 변경
